native: long -> s32 in LP64 signature mismatches (group A)
GC 'long' == 32-bit; on LP64 these decls no longer matched their s32/.cpp uses:
- CardSave changeMode/setMessage; JDRLighting setLightNum/setAmbNum;
  MapObjGrass entryGrassGroup; NpcBase setBalloonMessage;
  MapObjBase loadHideObjInfo (long*->s32*); ToolData GetValue (long&->s32&);
  CameraDemo startDemoCamera arg; dolphin/dvd.h callback first param (long->s32).

+++ include/GC2D/CardSave.hpp
@@ -17,7 +17,7 @@ class TPauseMenu2;
 
 class TCardSave : public JDrama::TViewObj {
 public:
-	TEProgress changeMode(long);
+	TEProgress changeMode(s32);
 
 	TCardSave(const char* name = "<TCardSave>", bool = false);
 
@@ -26,7 +26,7 @@ public:
 	void init(int);
 	void perform(u32, JDrama::TGraphics*);
 	void makeBuffer(J2DTextBox*, int);
-	void setMessage(J2DTextBox*, long, u32);
+	void setMessage(J2DTextBox*, s32, u32);
 	s8 waitForStop(TEProgress);
 	void endWaitForChoice();
 	s8 waitForChoice(TEProgress, TEProgress, signed char);
